chore: share calculation UI helpers
